refactor(v1): example envs put each server in its own self-launching servers/<name>.py
Separate server code from taskset code: each env's tool/user server moves out of the taskset
module into <env>/servers/<name>.py, a self-launching module ending with
`if __name__ == "__main__": <Server>.run()` (framework launches `python -m <env>.servers.<name>`).
The taskset module imports the server from .servers and uses it in tools()/user(); shared
constants/data the server needs live in the server module. Flat envs (glossary, deepwiki) become
packages; package envs drop their __main__.py.

- glossary -> servers/facts.py (+ facts.json beside it)
- deepwiki -> servers/deepwiki.py
- alphabet_sort, color_codeword -> servers/user.py
- wiki_search, wikispeedia -> servers/wiki.py (wikispeedia keeps graph.py in the package root)

GUIDE.md "Tools and user simulators" rewritten to the current vf-native surface (vf.Toolset /
vf.User classes, @vf.tool / respond, setup / setup_task, the servers/<name>.py layout, per-server
placement with own-host-runtime default; tools + users can coexist).

Verified: all 6 envs' server classes resolve to <env>.servers.<name>; glossary (tool) reward 1.0
on subprocess + docker; alphabet-sort (user sim) reward 1.0 on subprocess + docker.
