--- environments/alphabet_sort_v1/alphabet_sort_v1/__main__.py
@@ -1,5 +1,0 @@
-"""`python -m alphabet_sort_v1` — serve the user simulator (the framework's launch target)."""
-
-from alphabet_sort_v1 import AlphabetSortUser
-
-AlphabetSortUser.run()
