+++ environments/alphabet_sort_v1/alphabet_sort_v1/servers/user.py
@@ -1,0 +1,26 @@
+"""alphabet-sort user simulator: a `vf.User` that replays the episode's follow-up turns.
+
+Self-launching — the framework starts it with `python -m alphabet_sort_v1.servers.user`.
+"""
+
+import verifiers.v1 as vf
+
+
+class AlphabetSortUser(vf.User[vf.UserConfig]):
+    """Replays the episode's pre-generated follow-up turns: one `respond` per assistant turn,
+    injecting the next follow-up as a user message until all turns are done."""
+
+    async def setup_task(self, task) -> None:
+        self.follow_ups = task.info["follow_ups"]  # per-task input, from the task
+        self.num_turns = task.info["num_turns"]  # per-task input
+        self.turns = 0  # per-rollout mutable state
+
+    async def respond(self, message: str) -> tuple[vf.Messages, bool]:
+        self.turns += 1
+        if self.turns >= self.num_turns:
+            return [], True
+        return [{"role": "user", "content": self.follow_ups[self.turns - 1]}], False
+
+
+if __name__ == "__main__":
+    AlphabetSortUser.run()
