--- environments/color_codeword_v1/color_codeword_v1/__main__.py
@@ -1,5 +1,0 @@
-"""`python -m color_codeword_v1` — serve the user simulator (the framework's launch target)."""
-
-from color_codeword_v1 import ColorCodewordUser
-
-ColorCodewordUser.run()
