+++ environments/glossary_v1/pyproject.toml
@@ -9,6 +9,6 @@ dependencies = ["verifiers"]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
-[tool.hatch.build]
-include = ["glossary_v1.py", "facts.json", "pyproject.toml"]
+[tool.hatch.build.targets.wheel]
+packages = ["glossary_v1"]
