+++ environments/wiki_search_v1/wiki_search_v1/__init__.py
@@ -11,6 +11,8 @@ asks a judge model whether the harness's answer matches the ground truth.
 import verifiers.v1 as vf
 from verifiers.v1.dialects import ChatDialect
 
+from wiki_search_v1.servers.wiki import WikiSearchToolset
+
 SYSTEM = (
     "Use the Wikipedia search tools — `wiki_search_pages` to find relevant pages, "
     "`wiki_view_sections` to list a page's sections, and `wiki_read_section` to read "
@@ -38,9 +40,6 @@ Response:
 Respond either "yes" or "no" only. If a response contains incoherent text, respond \
 with "no" even if the correct answer is also present."""
 
-DATASET = "willcb/rare-wiki-pages"
-
-
 # The question bank and count are fixed properties of this env, not eval-time
 # knobs (the searchable corpus is built in `WikiSearchToolset.setup`).
 QUESTIONS_DATASET = "willcb/wiki-trivia-questions-v4"
@@ -63,91 +62,6 @@ class WikiSearchConfig(vf.TasksetConfig):
     # runtime), reused across rollouts rather than rebuilt per rollout. CLI-tunable, e.g.
     # `--taskset.tools.shared false` or `--taskset.tools.runtime.type docker`.
     tools: vf.ToolsetConfig = vf.ToolsetConfig(shared=True)
-
-
-class WikiSearchToolset(vf.Toolset[vf.ToolsetConfig]):
-    """Read-only search/view/read over the wiki corpus. The corpus + chroma index (expensive) are
-    built once in `setup`, in the server process. Every tool call is a read."""
-
-    TOOL_PREFIX = "wiki"  # the model sees `wiki_search_pages` / `wiki_view_sections` / `wiki_read_section`
-
-    async def setup(self) -> None:
-        import os
-        from pathlib import Path
-
-        # chromadb (via opentelemetry) imports protobuf-generated code predating protobuf 6.x;
-        # force the pure-Python impl so it imports next to a newer protobuf (e.g. vllm's).
-        os.environ.setdefault("PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION", "python")
-        import chromadb
-        from datasets import load_dataset
-
-        rows = load_dataset(DATASET, split="train")
-        self.pages = {r["id"]: {"title": r["title"], "content": r["content"]} for r in rows}
-        cache = os.environ.get("WIKI_SEARCH_CACHE", str(Path.home() / ".cache" / "wiki_search"))
-        col = chromadb.PersistentClient(path=f"{cache}/chroma").get_or_create_collection(
-            "wiki_titles"  # default local embedder (onnx MiniLM) — no API key
-        )
-        ids = list(self.pages)
-        have: set[str] = set()
-        for i in range(0, len(ids), 500):
-            have.update(col.get(ids=ids[i : i + 500]).get("ids", []))
-        missing = [pid for pid in ids if pid not in have]
-        for i in range(0, len(missing), 256):
-            batch = missing[i : i + 256]
-            col.upsert(ids=batch, documents=[self.pages[pid]["title"] for pid in batch])
-        self.collection = col
-
-    @staticmethod
-    def _norm(text: str) -> str:
-        """lowercase, spaces -> underscores (section ids)."""
-        return text.strip().lower().replace(" ", "_")
-
-    @vf.tool
-    def search_pages(self, query: str) -> list[dict]:
-        """Search the wiki for the 10 most relevant pages (by title). Returns
-        `[{page_id, title}, ...]` — pass a page_id to view_sections."""
-        result = self.collection.query(query_texts=[query], n_results=10)
-        return [
-            {"page_id": pid, "title": self.pages[pid]["title"]}
-            for pid in result["ids"][0]
-        ]
-
-    @vf.tool
-    def view_sections(self, page_id: str) -> list[dict]:
-        """List a page's sections. Returns `[{section_id, section_name}, ...]` — pass a
-        section_id to read_section."""
-        content = self.pages[page_id]["content"]
-        sections = [
-            {
-                "section_id": f"{page_id}:{self._norm(line.lstrip('#').strip())}",
-                "section_name": line.lstrip("#").strip(),
-            }
-            for line in content.split("\n")
-            if line.startswith("#")
-        ]
-        return sections or [
-            {"section_id": f"{page_id}:full", "section_name": "Full Page"}
-        ]
-
-    @vf.tool
-    def read_section(self, section_id: str) -> str:
-        """Read the text of a section (`page_id:section`; use `page_id:full` for all)."""
-        page_id, section = section_id.split(":", 1)
-        content = self.pages[page_id]["content"]
-        if section == "full":
-            return content
-        lines = content.split("\n")
-        start = end = None
-        for i, line in enumerate(lines):
-            if line.startswith("#"):
-                if self._norm(line.lstrip("#").strip()) == section and start is None:
-                    start = i
-                elif start is not None:
-                    end = i
-                    break
-        if start is None:
-            return f"Section not found: {section_id}"
-        return "\n".join(lines[start : end or len(lines)])
 
 
 class WikiSearchTaskset(vf.Taskset[TriviaTask, WikiSearchConfig]):
