--- environments/wiki_search_v1/wiki_search_v1/__main__.py
@@ -1,5 +1,0 @@
-"""`python -m wiki_search_v1` — serve the wiki-search tool server (the framework's launch target)."""
-
-from wiki_search_v1 import WikiSearchToolset
-
-WikiSearchToolset.run()
