--- environments/wikispeedia_v1/wikispeedia_v1/__main__.py
@@ -1,5 +1,0 @@
-"""`python -m wikispeedia_v1` — serve the wikispeedia tool server (the framework's launch target)."""
-
-from wikispeedia_v1 import WikiToolset
-
-WikiToolset.run()
